Refactor: ItemDecoration 클래스 분리 #25
